+++ src/components/PaginatedContainer/PaginatedContainer.scss
@@ -1,5 +1,7 @@
 .EventColumn {
   height: 100%;
   display: flex;
+  justify-content: center;
+  text-align: center;
   flex-direction: column;
 }
