+++ src/components/ProductCategory/ProductCategory.scss
@@ -54,6 +54,7 @@
   display: flex;
   flex-direction: column;
   margin: 0 10%;
+  height: 80vh;
 }
 
 .hide {
