+++ src/layout/EventTabs/Supply/index.js
@@ -1,16 +1,17 @@
 import React, { useState, useContext, useEffect } from "react";
 import PropTypes from "prop-types";
 
-import { userService } from "../../../Authentication/service";
+// import { userService } from "../../../Authentication/service";
 import { FormContext } from "../../../context/FormContext";
-// import { newSupplyContainerForm } from "./FormsToBeRendered/FormsToBeRendered";
 
 import "./Supply.scss";
+import createSetOfCategories from "./helper";
 
 import { Button } from "../../../components/Button";
-import { TextInput } from "../../../components/Inputs";
 import SupplyCategory from "../../../components/ProductCategory";
 import Modal from "../../../components/Modal";
+import PaginatedContainer from "../../../components/PaginatedContainer";
+import AddNewProductContainer from "./AddNewProductContainer";
 
 import products from "./Data/RealTempData";
 
@@ -22,41 +23,6 @@ const Supply = ({ id }) => {
   // const [supplyList, setsupply] = useState({
   //   SupplyContainers
   // });
-
-  const createSetOfCategories = array => {
-    const newArray = [];
-
-    const ListOfCategories = [];
-    array.forEach(catgory => {
-      ListOfCategories.push(catgory.productCategory.name);
-    });
-
-    const uniqCategory = [...new Set(ListOfCategories)];
-
-    uniqCategory.forEach(cate => {
-      const tempObject = { Category: cate, supplies: [] };
-
-      for (let i = 0; i < array.length; i += 1) {
-        let tempSupply = {};
-        if (cate === array[i].productCategory.name) {
-          array[i].userCards.forEach(user => {
-            tempSupply = {
-              id: user.eventUser.user.id,
-              supply: array[i].name,
-              userId: user.eventUser.user.id,
-              user: user.eventUser.user.username,
-              quantity: 2,
-              price: array[i].price,
-              picUrl: user.eventUser.user.picUrl
-            };
-          });
-          tempObject.supplies.push(tempSupply);
-        }
-      }
-      newArray.push(tempObject);
-    });
-    return newArray;
-  };
 
   useEffect(() => {
     // console.log("mounted");
@@ -78,89 +44,54 @@ const Supply = ({ id }) => {
     //   .catch(err => {
     //     console.log(err);
     //   });
-    return () => {
-      console.log("unmounted");
-    };
+    return () => {};
   }, [productsTemp]);
 
   const [forms, setform] = useContext(FormContext);
 
-  const [newEvent, setNewEvent] = useState({
-    name: "",
-    price: "",
-    productCategory: ""
-  });
-
-  const openModalNewSupplyContainer = () => {
-    setform({ show: true });
-  };
-  const onChangeHandler = event => {
-    setNewEvent({ ...newEvent, [`${event.target.name}`]: event.target.value });
-    console.log(newEvent);
+  const addNewProductModal = () => {
+    setform({ renderForm: <AddNewProductContainer />, show: true });
   };
 
-  const addNewProduct = event => {
-    event.preventDefault();
-    userService
-      .addNewSuplyToEvent(id, newEvent)
-      .then(body => {
-        return body;
-      })
-      .then(res => {
-        console.log(res);
-      })
-      .catch(err => {
-        console.log(err);
-      });
-  };
+  // const addNewProduct = event => {
+  //   event.preventDefault();
+  //   userService
+  //     .addNewSuplyToEvent(id, newEvent)
+  //     .then(body => {
+  //       return body;
+  //     })
+  //     .then(res => {
+  //       console.log(res);
+  //     })
+  //     .catch(err => {
+  //       console.log(err);
+  //     });
+  // };
 
   const hideModal = () => {
     setform({ ...forms, show: false });
   };
 
-  const newSupplyContainerForm = () => {
-    return (
-      <div>
-        <TextInput
-          onChange={onChangeHandler}
-          placeholder="Supply name"
-          name="name"
-        />
-        <TextInput
-          onChange={onChangeHandler}
-          type="number"
-          placeholder="Price"
-          name="price"
-        />
-        <TextInput
-          onChange={onChangeHandler}
-          placeholder="Supply Container Name"
-          name="productCategory"
-        />
-        <Button clicked={addNewProduct} classes="btn-blueGradient btn-md">
-          apply
-        </Button>
-      </div>
-    );
-  };
-
   return (
     <div className="SuplyBody">
       <Modal show={forms.show} modalClose={hideModal}>
-        {newSupplyContainerForm()}
+        {forms.renderForm}
       </Modal>
       <div className="buttonContainer">
-        <Button
-          classes="btn-md btn-blueGradient"
-          clicked={openModalNewSupplyContainer}
-        >
+        <Button classes="btn-md btn-blueGradient" clicked={addNewProductModal}>
           Add new supply +
         </Button>
       </div>
-
-      {supplyList2.map(supCont => {
-        return <SupplyCategory supCont={supCont} key={supCont.Category} />;
-      })}
+      <PaginatedContainer
+        title="Product list"
+        items={supplyList2}
+        perPage={5}
+        render={({ items }) =>
+          items.map(supCont => (
+            <SupplyCategory supCont={supCont} key={supCont.Category} />
+          ))
+        }
+      />
     </div>
   );
 };
